agents/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-openai>=0.10.0
livekit-plugins-google>=0.10.0
python-dotenv>=1.0.0
